Extract counter component
